DB 생성 작업까지 완성

+++ src/main/java/com/project/sroa/model/UserInfo.java
@@ -16,5 +16,5 @@ public class UserInfo {
     public String name;
     public String address;
     public long phoneNum;
-    public long code;
+    public Integer code;
 }
